Add a function to preprocess the bed file
To preprocess the bed file pass in arguments like:

--preprocess --bfile ./data/uk10k_chr1_1mb --pheno ./data/test.phen

This will output a .ppbed file containing the calculated values
with a mean of 0 and a sqnorm file with the square normalised
values in them.

These files are conveniently in the form that we can then mmap
the files and present them by way of mapped matrices with Eigen.

On my local machine it takes 0.2s to preprocess the 5.8MiB test
bed file. The preprocessed file is much larger than the original
bed file as we are converting to a float each pair of bits stored
in the original. As long as the filesystem supports large files
though this will be fine.