Implement a task using the mapped data files
* Use the mapped preprocessed data files.
* These are accessed via Eigen Map helpers but otherwise the
  pattern is the same as for regular Eigen matrices/vectors.
* Use std::iota to generate the vector of indices.
* Print out progress every 1000'th iteration rather than when
  the sample number happens to be a multiple of 1000.
* Add progress indicator to output - useful for the larger
  datasets.

With this approach the preprocessed data file for the

ukb_imp_v3_UKB_EST_clumpLD09_chr22.ppbed

dataset is 7.5GB, yet when processing the memory usage of the
application never exceeds 40MB. With an SSD the performance seems
very close to that of holding everything in physical memory for
the smaller datasets.

Now we have the advantage of being able to process essentially
unbounded dataset sizes (limited by disk).

The preprocessing approach seems very much worth it compared with
the alternative approach of trying to calculate data upon access
as has been attempted on the main branch. We only do the work for
each marker exactly once whereas the other approach must do it
every time that marker is selected at random.

Thinking of this random access pattern, the performance may
benefit by telling the kernel about it. We can do this via the
madvise() system call using the MADV_RANDOM value. This tells
the kernel to not bother as much with pre-loading the next
few pages into physical memory. A sequential access pattern
benefits from this of course, but for our random access it
is likely just wasted work for the kernel and IO subsystem.

This approach therefore reduced the computational expense at
runtime. We also still have the option to employ some multi-
threading scheme to parallelise the actual analysis. But that
is a different problem.