Fix crash on larger datasets
The issue is that when calculating the length of the region to
mmap, the multiplication was overflowing the unsigned int
temporary variable (as the number of snps and individuals are
stored in unsigned ints). Fix this by explicitly converting these
to size_t in the calculations.

Now accessing the mappedZ variable works for any column as we
are actually mapping the entire file now.